edit solution for 2-

+++ 0x15-file_io/2-append_text_to_file.c
@@ -10,7 +10,7 @@
 int append_text_to_file(const char *filename, char *text_content)
 {
 	int fd, i;
-	size_t writeByte;
+	int writeByte;
 
 	if (filename == NULL)
 		return (-1);
